+++ UnderlyingPath.cu
@@ -4,12 +4,12 @@
 #include "RandomGenerator.h"
 #include "StocasticProcess.h"
 
-__host__ __device__  MontecarloPath::MontecarloPath(float SInitial, float TInitial, float TFinal,RandomGenerator* Generator,StocasticProcess* Process, int NSteps){
+__host__ __device__  MontecarloPath::MontecarloPath(double SInitial, double TInitial, double TFinal,RandomGenerator* Generator,StocasticProcess* Process, int NSteps){
     _TInitial = TInitial;
     _SInitial = SInitial;
     _Generator = Generator;
     _Process = Process;
-    _UnderlyingPath = new float[NSteps];
+    _UnderlyingPath = new double[NSteps];
     _TFinal=TFinal;
     _NSteps=NSteps;
 };
@@ -18,17 +18,17 @@ __host__ __device__  MontecarloPath::~MontecarloPath(){
     delete[] _UnderlyingPath;
 };
 
-__host__ __device__  float* MontecarloPath::GetPath(){
+__host__ __device__  double* MontecarloPath::GetPath(){
 
-    float TStep = (_TFinal-_TInitial)/_NSteps;
-    float temp=_SInitial;
+    double TStep = (_TFinal-_TInitial)/_NSteps;
+    double temp=_SInitial;
 
     for(int i=0; i<_NSteps; i++){
-        float w = _Generator->Gauss();
+        double w = _Generator->Gauss();
         temp = _Process->Step(temp, TStep, w);
         _UnderlyingPath[i]=temp;
     }
-    
+
     return _UnderlyingPath;
 
 };
